+++ wear/FindMyPhoneSample/Wearable/MainActivity.cs
@@ -41,7 +41,7 @@ namespace FindMyPhoneSample
 					.SetContentAction (0)
 					.SetHintHideIcon (true))
 				.SetLocalOnly (true)
-				.SetPriority ((int)NotificationPriority.Max);
+				.SetPriority (NotificationPriority.Max);
 			((NotificationManager)GetSystemService (NotificationService))
 				.Notify (FIND_PHONE_NOTIFICATION_ID, notification.Build ());
